# openapi-schema.json
openapi: 3.0.3
info:
  title: ChargeGhar Nepal API
  version: a315307-dirty
  description: ChargeGhar API - Shared Power Bank Network for Nepal. This documentation
    only shows endpoints documented in our Features TOC.
  contact:
    name: ChargeGhar API Support
    email: support@chargegh.com
  license:
    name: Proprietary
paths:
...
  /api/rentals/{rental_id}/cancel:
    post:
      operationId: api_rentals_cancel_create
      description: Cancel an active rental with optional reason
      summary: Cancel Active Rental
      parameters:
      - in: path
        name: rental_id
        schema:
          type: string
        required: true
      tags:
      - Rentals
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/RentalCancelRequest'
          application/json:
            schema:
              $ref: '#/components/schemas/RentalCancelRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/RentalCancelRequest'
      security:
      - cookieAuth: []
      - basicAuth: []
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BaseResponse'
          description: ''
  /api/rentals/{rental_id}/extend:
    post:
      operationId: api_rentals_extend_create
      description: Extend rental duration by purchasing additional time package
      summary: Extend Rental Duration
      parameters:
      - in: path
        name: rental_id
        schema:
          type: string
        required: true
      tags:
      - Rentals
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/RentalExtensionCreateRequest'
          application/json:
            schema:
              $ref: '#/components/schemas/RentalExtensionCreateRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/RentalExtensionCreateRequest'
        required: true
      security:
      - cookieAuth: []
      - basicAuth: []
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BaseResponse'
          description: ''
  /api/rentals/{rental_id}/issues:
    post:
      operationId: api_rentals_issues_create
      description: Report an issue with current rental
      summary: Report Rental Issue
      parameters:
      - in: path
        name: rental_id
        schema:
          type: string
        required: true
      tags:
      - Rentals
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/RentalIssueCreateRequest'
          application/json:
            schema:
              $ref: '#/components/schemas/RentalIssueCreateRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/RentalIssueCreateRequest'
        required: true
      security:
      - cookieAuth: []
      - basicAuth: []
      - jwtAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BaseResponse'
          description: ''
  /api/rentals/{rental_id}/location:
    post:
      operationId: api_rentals_location_create
      description: Update GPS location for active rental tracking
      summary: Update Rental Location
      parameters:
      - in: path
        name: rental_id
        schema:
          type: string
        required: true
      tags:
      - Rentals
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/RentalLocationUpdateRequest'
          application/json:
            schema:
              $ref: '#/components/schemas/RentalLocationUpdateRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/RentalLocationUpdateRequest'
        required: true
      security:
      - cookieAuth: []
      - basicAuth: []
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BaseResponse'
          description: ''
  /api/rentals/{rental_id}/pay-due:
    post:
      operationId: api_rentals_pay_due_create
      description: Settle outstanding rental dues using points and wallet combination
      summary: Settle Rental Dues
      parameters:
      - in: path
        name: rental_id
        schema:
          type: string
        required: true
      tags:
      - Rentals
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/RentalPayDueRequest'
          application/json:
            schema:
              $ref: '#/components/schemas/RentalPayDueRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/RentalPayDueRequest'
        required: true
      security:
      - cookieAuth: []
      - basicAuth: []
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BaseResponse'
          description: ''
  /api/rentals/active:
    get:
      operationId: api_rentals_active_retrieve
      description: Returns user's current active rental if any
      summary: Get Active Rental
      tags:
      - Rentals
      security:
      - cookieAuth: []
      - basicAuth: []
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BaseResponse'
          description: ''
  /api/rentals/history:
    get:
      operationId: api_rentals_history_retrieve
      description: Retrieve user's rental history with optional filtering
      summary: Get Rental History
      parameters:
      - in: query
        name: end_date
        schema:
          type: string
          format: date
        description: Filter rentals until this date
      - in: query
        name: page
        schema:
          type: integer
        description: Page number for pagination
      - in: query
        name: page_size
        schema:
          type: integer
        description: Number of items per page
      - in: query
        name: payment_status
        schema:
          type: string
        description: Filter by payment status
      - in: query
        name: start_date
        schema:
          type: string
          format: date
        description: Filter rentals from this date
      - in: query
        name: status
        schema:
          type: string
        description: Filter by rental status
      tags:
      - Rentals
      security:
      - cookieAuth: []
      - basicAuth: []
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BaseResponse'
          description: ''
  /api/rentals/packages:
    get:
      operationId: api_rentals_packages_retrieve
      description: Get list of available rental packages with pagination
      summary: Get Rental Packages
      parameters:
      - in: query
        name: page
        schema:
          type: integer
        description: Page number
      - in: query
        name: page_size
        schema:
          type: integer
        description: Items per page
      tags:
      - Rentals
      security:
      - cookieAuth: []
      - basicAuth: []
      - jwtAuth: []
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BaseResponse'
          description: ''
  /api/rentals/start:
    post:
      operationId: api_rentals_start_create
      description: Start a new power bank rental at specified station with selected
        package
      summary: Start New Rental
      tags:
      - Rentals
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/RentalStartRequest'
          application/json:
            schema:
              $ref: '#/components/schemas/RentalStartRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/RentalStartRequest'
        required: true
      security:
      - cookieAuth: []
      - basicAuth: []
      - jwtAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BaseResponse'
          description: ''
  /api/rentals/stats:
    get:
      operationId: api_rentals_stats_retrieve
      description: Get comprehensive rental statistics for the user
      summary: Get Rental Statistics
      tags:
      - Rentals
      security:
      - cookieAuth: []
      - basicAuth: []
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BaseResponse'
          description: ''
...
components:
  schemas:
...
    BaseResponse:
      type: object
      description: Standard response format for all API endpoints
      properties:
        success:
          type: boolean
        message:
          type: string
        data: {}
        timestamp:
          type: string
          format: date-time
      required:
      - data
      - message
      - success
      - timestamp
...
    RentalCancelRequest:
      type: object
      description: |-
        Request serializer for rental cancellation.
        Used in: POST /api/rentals/{id}/cancel
      properties:
        reason:
          type: string
          description: Optional cancellation reason
          maxLength: 255
    RentalExtensionCreateRequest:
      type: object
      description: |-
        Request serializer for extending a rental.
        Used in: POST /api/rentals/{id}/extend
      properties:
        package_id:
          type: string
          format: uuid
          description: Extension package ID
      required:
      - package_id
    RentalIssueCreateRequest:
      type: object
      description: |-
        Request serializer for reporting rental issues.
        Used in: POST /api/rentals/{id}/issue
      properties:
        issue_type:
          enum:
          - POWER_BANK_DAMAGED
          - POWER_BANK_LOST
          - CHARGING_ISSUE
          - RETURN_ISSUE
          type: string
          description: |-
            * `POWER_BANK_DAMAGED` - Power Bank Damaged
            * `POWER_BANK_LOST` - Power Bank Lost
            * `CHARGING_ISSUE` - Charging Issue
            * `RETURN_ISSUE` - Return Issue
          x-spec-enum-id: 9d97377ca8c5f924
        description:
          type: string
          minLength: 1
          maxLength: 255
        images: {}
      required:
      - description
      - issue_type
    RentalLocationUpdateRequest:
      type: object
      description: |-
        Request serializer for updating rental location.
        Used in: POST /api/rentals/{id}/location
      properties:
        latitude:
          type: number
          format: double
          description: GPS latitude (-90 to 90)
        longitude:
          type: number
          format: double
          description: GPS longitude (-180 to 180)
        accuracy:
          type: number
          format: double
          default: 10.0
          description: GPS accuracy in meters
      required:
      - latitude
      - longitude
...
    RentalPayDueRequest:
      type: object
      description: |-
        Request serializer for paying rental dues.
        Used in: POST /api/rentals/pay-due
        Note: Same structure as calculate-options for settle_dues scenario
      properties:
        scenario:
          enum:
          - settle_dues
          type: string
          x-spec-enum-id: dce4e640f37717d2
          description: |-
            Payment scenario (only settle_dues supported)

            * `settle_dues` - settle_dues
        package_id:
          type: string
          format: uuid
          description: Package ID for settlement
        rental_id:
          type: string
          format: uuid
          description: Rental ID with outstanding dues
      required:
      - package_id
      - rental_id
      - scenario
    RentalStartRequest:
      type: object
      description: |-
        Request serializer for starting a rental.
        Used in: POST /api/rentals/start
      properties:
        station_sn:
          type: string
          minLength: 1
          description: Station serial number where powerbank is picked up
          maxLength: 255
        package_id:
          type: string
          format: uuid
          description: Selected rental package ID
        payment_scenario:
          enum:
          - pre_payment
          - post_payment
          type: string
          x-spec-enum-id: 96b65b92ff822ca5
          description: |-
            Payment timing preference

            * `pre_payment` - pre_payment
            * `post_payment` - post_payment
      required:
      - package_id
      - payment_scenario
      - station_sn
...
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
    cookieAuth:
      type: apiKey
      in: cookie
      name: sessionid
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...